mg_miner/requirements.txt
flake8
black
pytest
mypy
bandit
selenium
behave
locust
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
grafanalib
retrying

